+++ apps/be/src/main/java/com/capstone/logue/anal/dto/fastapi/request/DataSourceMeta.java
@@ -1,8 +1,6 @@
-package com.capstone.logue.anal.dto.request;
+package com.capstone.logue.anal.dto.fastapi.request;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
-import lombok.Builder;
-import lombok.Getter;
 
 import java.util.List;
