+++ apps/be/src/test/java/com/capstone/logue/anal/service/AnalServiceTest.java
@@ -8,11 +8,10 @@ import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
 import com.capstone.logue.anal.client.FastApiClient;
-import com.capstone.logue.anal.dto.request.CreateAnalysisFlowRequest;
-import com.capstone.logue.anal.dto.response.CreateAnalysisFlowResponse;
-import com.capstone.logue.anal.dto.response.CreateConversationResponse;
+import com.capstone.logue.anal.dto.spring.request.CreateAnalysisFlowRequest;
+import com.capstone.logue.anal.dto.spring.response.CreateAnalysisFlowResponse;
+import com.capstone.logue.anal.dto.spring.response.CreateConversationResponse;
 import com.capstone.logue.anal.repository.*;
-import com.capstone.logue.auth.security.UserPrincipal;
 import com.capstone.logue.data.repository.DataSourceRepository;
 import com.capstone.logue.auth.provider.SecurityContextProvider;
 import com.capstone.logue.global.entity.*;
@@ -21,8 +20,6 @@ import com.capstone.logue.global.entity.enums.JobStatus;
 import com.capstone.logue.global.exception.ErrorCode;
 import com.capstone.logue.global.exception.LogueException;
 import com.capstone.logue.user.repository.UserRepository;
-import com.capstone.logue.anal.dto.fastapi.FileAnalysisResponse;
-import com.capstone.logue.anal.dto.request.FileAnalysisRequest;
 import com.fasterxml.jackson.databind.node.ObjectNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.junit.jupiter.api.BeforeEach;
@@ -32,8 +29,6 @@ import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 import org.springframework.web.client.RestTemplate;
-import org.springframework.http.HttpMethod;
-import org.springframework.http.ResponseEntity;
 import org.springframework.test.util.ReflectionTestUtils;
 
 import java.time.OffsetDateTime;
